pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
python-dotenv = "^0.14.0"
requests = "^2.24.0"
pytest = "^6.1.1"
pyparsing = "^2.4.7"
Werkzeug = "^1.0.1"
wcwidth = "^0.2.5"
urllib3 = "^1.25.11"
selenium = "^3.141.0"
gunicorn = "^20.0.4"
pymongo = {extras = ["srv"], version = "^3.11.4"}
mongomock = "^3.23.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"